Get BASE data variable availability

This function obtains the BASE data availability for all or selected AmeriFlux sites. See AmeriFlux page https://ameriflux.lbl.gov/data/aboutdata/data-variables/ for details about the variable naming.

amf_list_data(site_set = NULL, var_set = NULL)

Arguments

site_set

A scalar or vector of character specifying the target AmeriFlux Site ID (CC-Sss). If not specified, it returns all sites.

var_set

A scalar or vector of character specifying the target variables as in basename. See AmeriFlux pagehttps://ameriflux.lbl.gov/data/aboutdata/data-variables/#base for a list of variable names. If not specified, it returns all variables.

Value

A data frame of variable-specific data availability (per year) for selected AmeriFlux sites.

Site_ID - Six character site identifier (CC-Sss)

VARIABLE - Variable name of the data included in the BASE file

BASENAME - Variable base name of the data included in the BASE file.

GAP_FILLED - Whether a variable is a gap-filled variable (TRUE/FALSE)

Y1990 - Percentage of data availability in the year 1990 (0-1).

Y1991 - Percentage of data availability in the year 1991 (0-1).

Y1992 - Percentage of data availability in the year 1992 (0-1).

...

Examples

if (FALSE) {
# obtain the data variable availability for all sites
data_aval <- amf_list_data()

# obtain the data variable availability for selected sites
data_aval <- amf_list_data(site_set = c("US-CRT","US-WPT"))

# obtain the data variable availability for selected variables
data_aval <- amf_list_data(var_set = c("FCH4", "WTD"))

}