Get variable information

This function obtains the measurement height metadata for the AmeriFlux BASE data product. See AmeriFlux page https://ameriflux.lbl.gov/data/measurement-height/ for details.

amf_var_info(out_dir = tempdir(), verbose = TRUE)

Arguments

out_dir
output directory (default = tempdir())

verbose
Show feedback on download progress (TRUE/FALSE)

Value

A data frame of measurement height data for all avaiable AmeriFlux sites #'

Site_ID - Six character site identifier (CC-Sss)

Variable - Variable name of the data included in the BASE file

Start_Date - Date when the information first applies

Height - Distance above the ground surface in meters

Instrument_Model - Instrument model used to collect the data variable

Instrument_Model2 - A second instrument model used to collect the data variable

Comment - Additional information provided by the site team

BASE_Version - The most recent BASE data product version number for which the information applies

Examples

if (FALSE) {
# download the measurement height data for all sites
var_info <- amf_var_info()
}